Bytecode: enforce consistent typing of size variables

+++ mex/sources/bytecode/Interpreter.hh
@@ -53,7 +53,7 @@ protected:
 public:
   Interpreter(Evaluate &evaluator_arg, double *params_arg, double *y_arg, double *ya_arg, double *x_arg, double *steady_y_arg,
               double *direction_arg, int y_size_arg,
-              size_t nb_row_x_arg, int periods_arg, int y_kmin_arg, int y_kmax_arg,
+              int nb_row_x_arg, int periods_arg, int y_kmin_arg, int y_kmax_arg,
               int maxit_arg_, double solve_tolf_arg, int y_decal_arg, double markowitz_c_arg,
               string &filename_arg, int minimal_solving_periods_arg, int stack_solve_algo_arg, int solve_algo_arg,
               bool global_temporary_terms_arg, bool print_arg, mxArray *GlobalTemporaryTerms_arg,
